requirements.txt
asgiref==3.6.0
autopep8==2.0.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.10.0
firebase-admin==6.1.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.21.0
idna==3.4
msgpack==1.0.5
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.1
platformdirs==3.1.1
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
xlrd==2.0.1
XlsxWriter==3.1.1
